+++ src/main/java/faang/school/analytics/listener/AbstractEventListener.java
@@ -1,7 +1,6 @@
 package faang.school.analytics.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.analytics.model.EventType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
@@ -13,8 +12,6 @@ import java.util.function.Consumer;
 @RequiredArgsConstructor
 public abstract class AbstractEventListener<T> implements MessageListener {
     private final ObjectMapper objectMapper;
-
-    public abstract EventType getEventType();
 
     protected void handleEvent(Message message, Class<T> tClass, Consumer<T> consumer) {
         try {
